Build the root level version in a subdir and move
Building the version deployed to the root of the bucket caused an exception within Jekyll regarding stack level during the cleanup tasks.

We now build the root level version inside the `/site/root` folder and move it to the `/site` folder afterwards.

We also build the version after the version for the `/site/documentation` subfolder.